feat: Update and add new weapon configs  ## What Was Changed
 ### Replaced

- [40] Fast Shock Rocket -> Better Ad Vic (REQ 5)
- [23] H1 Mangler -> Burst Mangler (REQ 8)

 ### Added

- [49] Napalm Hydra (REQ 8)
- [50] Harbinger Weapon (REQ 9)

 ## Reasoning

 ### Better Ad Vic

The Fast Shock Rocket was deemed not worthy of a REQ 8 due to its Disruptor projectile gimmick not being enough as the weapon still was a 2-shot point-blank rocket kill. Attempted to double the rocket count per shot, but it resulted in not getting kill credit almost ever.

Replaced with "Better Ad Vic" which is a stronger version of the [39] Ad Victoriam. Fires three M41 SPNKr rockets instead of one M41 SPNKr rocket and two Rocket Hog rockets.

 ### H1 Mangler

The H1 Mangler was too similar to the Stalker Rifle Ultra at the same REQ tier and was actually performing worse than the Stalker Rifle Ultra, so it was cut.

Replaced with "Burst Mangler" which fires three shots per trigger pull with an immense recoil that is not seen on other weapons. Landing two body shots and a headshot on an enemy results in a kill, but this is very difficult to achieve outside of close range. The gun is very rewarding if you can control the recoil. Long range recoil control is nearly impossible.

 ### Napalm Hydra

Hydra that shoots Ravager Rebound shots and should be used with the secondary fire as they are charged shots that usually take up a while to charge up with the default Ravager Rebound. REQ 8 because these charged shots deal a lot of area-of-effect damage and 6 of them can be fired in quick succession.

 ### Harbinger Weapon

An unique weapon that we want to try and fit in the sandbox as the only other AI weapon (next to the Brute Scrap Cannon) that has working 3rd person animations, clamber and can be dropped without deletion.

Features:
- Fires blue tracking shots that stay in place for a while and then track to the nearest enemy, even through walls
- Multiple shots can be fired after each other semi-auto
- Has no first-person weapon model.

While it doesn't have a first-person model, we think it doesn't take away from the gameplay experience. We predict that players will care more about exploring a completely different weapon than to pay attention to the missing first-person weapon model.