feat: Update and add new weapon configs  ## Changes
**Replaced**
- [40] Fast Shock Rocket -> Better Ad Vic (REQ 5)
- [23] H1 Mangler -> Burst Mangler (REQ 8)

 **Added**
- [49] Napalm Hydra (REQ 8)
- [50] Harbinger Weapon (REQ 9)

 ## Reasoning

 ### Better Ad Vic

The Fast Shock Rocket was deemed not worthy of a REQ 8 due to its Disruptor projectile gimmick not being enough as the weapon still was a 2-shot point-blank rocket kill. Attempted to double the rocket count per shot, but it resulted in not getting kill credit almost ever.

Replaced with "Better Ad Vic":
- Stronger version of the [39] Ad Victoriam.
- Fires three M41 SPNKr rockets instead of one M41 SPNKr rocket and two Rocket Hog rockets.

 ### H1 Mangler

The H1 Mangler was too similar to the Stalker Rifle Ultra at the same REQ tier and was actually performing worse than the Stalker Rifle Ultra, so it was cut.

Replaced with "Burst Mangler":
- Fires three shots per trigger pull with an immense recoil that is not seen on other weapons.
- Kill with two body shots and a headshot; very difficult to achieve outside of close range.
- Rewarding if you can control the recoil.
- Long range recoil control nearly impossible.

 ### Napalm Hydra

- Hydra that shoots Ravager Rebound shots.
- Should be used with the secondary fire as it fires charged shots that usually take up a while to charge up with the default Ravager Rebound.
- REQ 8 because these charged shots deal a lot of area-of-effect damage and 6 of them can be fired in quick succession.

 ### Harbinger Weapon

An unique weapon that we want to try and fit in the sandbox as the only other AI weapon (next to the Brute Scrap Cannon) that has working 3rd person animations, clamber and can be dropped without deletion.

Features:
- Fires blue tracking shots that stay in place for a while and then track to the nearest enemy, even through walls
- Multiple shots can be fired after each other semi-auto
- Has no first-person weapon model.

While it doesn't have a first-person model, we think it doesn't take away from the gameplay experience. We predict that players will care more about exploring a completely different weapon than to pay attention to the missing first-person weapon model.