feat: Remove [50] and relocate [51] to [32]
 ## Changes

**Removed**
- [50] Harbinger Weapon, REQ 9

**Adjusted**
- [51] Brute Scrap Cannon, REQ 2 -> [32] Brute Scrap Cannon

 ## Reasoning

**Harbinger Weapon removal**
The Harbinger's Weapon which was pitched as a REQ 9 weapon in https://github.com/The-Scripters-Guild/Warzone/pull/20 for it being a potential viable weapon turned out to have an issue where the shot projectiles would prematurely explode before reaching the target, rendering the weapon useless and a waste of REQ points.

**Brute Scrap Cannon change**
As the [32] BurstCannon was removed in https://github.com/The-Scripters-Guild/Warzone/pull/21 and now the [50] Harbinger Weapon removed, it left the [51] Brute Scrap Cannon hanging off from the rest of the designated numbers, so it was relocated to 32.
This early in development this change should not have any negative impact on existing scripting, as long as the frameworks are updated to account for the number change in due course.