docs: armor-mod-config is severely behind. The dev names we were using for the armor mods are no longer needed. They have been updated to their new names we made using the "string" words that we have access to in forge.
We came to the conclusion of what game states the armor mods were going to be unlocked. This has been updated as well.

Upgraded Walking has now been fully tested and so it was required to be added. While Upgraded Walking gives you a passive boost to your strafe and or walking speed, Upgraded Sprinting is focused on giving you a passive boost to your actual sprint speed. These two armor mods function as inverse of each other and come with other added benfits that the other does not receive.