feat: Weapons need VFX. Add them. We had a conversation of what criteria we will be following for when a weapon should be granted a VFX. This is what became of that:
Vip: The user of the weapon its applied to is a threat but not an issue that should be considered a priority for more than 2 players.

Beta Infected: The weapon its applied to is a big enough threat that no more than 4 players should consider it a prioriy. Even if its in the hands of a inexperienced player

Infected alpha: A threat to a whole team, even if the weapon is in the hands of a inexperienced player.

Using this criteria these weapons have been given a VFX

[20] Attack Off Iratus: Beta Infected
[40] Guard Off Doisac: Beta Infected
[27] Scout Skewer: VIP