feat: Update weapon configs after PvP testing  ## Changes
**Added**
- [32] Spartan Sandwich (REQ 8)

**Removed**
- [37] Spike Hydra (REQ 5)
- [42] Plasma Mangler (REQ 6)

**Adjusted:**
- [48] Headhunter ammo adjustment 34 -> -70 (6+18 -> 6+6 shots)
- [36] Power Off Jega Rdomnai camouflage Intensity Scalar 0.80 -> 1.00 (full invisibility)
- [47] Scorpion Shot ammo adjustment -4.80 -> -3.5 (8 -> 6 shots)

**Chores:**
- Updated Progressive Purchases table to reflect weapon configuration changes.

 ## Spartan Sandwich addition

The Spartan Sandwich is another meme weapon alongside the [41] Go At that we thought to include as the Sandwich has become more popular as a funny weapon in the Halo Infinite community. We made it a high REQ and let it be a unique unlimited melee weapon with an increased movement speed and a red VFX indicating that it has a high threat level.

 ## Spike Hydra and Plasma Mangler removal

The Spike Hydra and Plasma Mangler were removed mainly because of moving to a more efficient projectile storage method that had trouble with getting the grenade-throwing AI to throw their grenade as the AI were not standing on Nav Mesh data anymore.
The weapons had subpar kill credit consistency as well due to the arcing projectile and the designers agreed that removing them from the sandbox wouldn't be a big loss for the sandbox variety. Multiple cloned grenades have also been the only projectiles that have proven to cause server-freezing crashes in testing, so that possibility is now avoided as well.

 ## Headhunter ammo

Following testing in a PvP environment using [Warzone Fiesta](https://okom.one/blog/warzone-fiesta-released), it was noted that the [48] Headhunter in it's secondary fire mode allowed for too high of a maximum kill potential from a medium range that wasn't accounted for. The weapon being able to one-hit kill enemies with a headshot at such long ranges while having 24 shots to spend was too much for a gun in it's REQ 7 tier when comparing to other weapons in the REQ tier such as the [13] Valkyrie and [45] Infiltrator Off Worlds.
The ammo was adjusted from 24 -> 12 shots to limit it's maximum kill potential while still keeping the weapon a good alternative for other weapons in it's REQ tier at a short–medium range.

 ## Power Off Jega Rdomnai camouflage

Following testing in a PvP environment using [Warzone Fiesta](https://okom.one/blog/warzone-fiesta-released), it was noted that the camouflage of the [36] Power Off Jega Rdomnai was too easily noticeable; having a dark tint in the camouflage. This tint in addition to the loud sound the Infected Energy Sword base weapon emits made the wielder too easy to spot, resulting in the REQ 6 weapon being subpar in it's unique effectiveness.
The camouflage Intensity Scalar was increased from 0.80 -> 1.00 to grant full invisibility to the wielder while still keeping the loud sounds present that emanate from the sword.

 ## Scorpion Shot ammo

Following testing in a PvP environment using [Warzone Fiesta](https://okom.one/blog/warzone-fiesta-released) and in a PvE environment using [Darkstar Firefight](https://www.halowaypoint.com/halo-infinite/ugc/maps/a10e9ab8-9ae3-4db8-a755-b136fa2ea1b7), it was noted that the ammo capacity of the [47] Scorpion Shot was too much in relation to how much potential damage the weapon can cause. In comparison, the other REQ 9 weapon, [43] Banish Off Balaho has 10 Volatile Skewer shots that do less damage and have much less area-of-effect damage than the Scorpion Shot. While the Scorpion Shot is limited by it's turret debuffs like movement speed, it was still too powerful with 8 shots.
The ammo was adjusted from 8 -> 6 shots to limit it's effectiveness and maximum kill potential while still keeping the weapon a good investment at REQ 9.