Go At base weapon change to Fuel Rod SPNKr Weapon Type: M41 SPNKr + S7 Flexfire -> Fuel Rod SPNKr + S7 Flexfire
The Weapon Type of the Go At (Agoat) was adjusted to be a Fuel Rod SPNKr + S7 Flexfire as we wanted a third variant of the Fuel Rod SPNKr and this change also keeps the SPNKr variants consistent with the logic that the Fuel Rod SPNKr base weapons have >2 rockets in the chamber, while the M41 SPNKr rockets do not.

The more muted sound of the Fuel Rod SPNKr's fire is also more pleasant for this weapon than the M41 SPNKr's, which makes the weapon sound more like it's inspiration of the Quake rocket launcher.