Replace [38] and remove [45]
Replace [38]: This weapon was our attempt at recreating the Fuel Rod Cannon from previous Halo titles, but now that the Fuel Rod SPNKr has been added to the game, this weapon can be remade with better UX.

The weapon was renamed to "Infiltrator Off Worlds", which was the upgraded Fuel Rod Cannon weapon recreation's name earlier, and here it sounds better than "Hunters Arm".

The Weapon Type was changed to have the Fuel Rod SPNKr instead of the Cindershot, but the weapon will still shoot a blank projectile so we can have a custom projectile.

The custom projectile was changed from a green Hunter Shot to a red Banished Hunter Shot to communicate a difference in the damage output. As the regular Fuel Rod SPNKr now outperforms what we had initially ranked this Hunters Arm weapon as, the green Fuel Rod projectile will now communicate a larger amount of damage than the red Banished Hunter Shot, which is the opposite of what we had up to this point.

Remove [45]: This weapon was an upgraded variant of the custom Hunter cannon weapon "[38] Hunters Arm" we made, but with the release of the Fuel Rod SPNKr, this weapon became obsolete as the Fuel Rod SPNKr serves nearly the same purpose at an official UX capacity.