feat: 38 new weapon type and ammo adjustment - Weapon Type: Fuel Rod SPNKr + Convergence Bulldog -> Fuel Rod SPNKr + Striker Sidekick - Ammo Adjustment: -42 (10+0) -> -13 (10+0)
After receiving a new node in the [Spring Update 2025](https://support.halowaypoint.com/hc/en-us/articles/34335369079700-Halo-Infinite-Spring-Update-2025-Patch-Notes#h_01JJQBZ4QWFM2D79NK0MQKSD39) for attributing scripted damage to a source object, we were able to majorly improve the kill credit for the Infiltrator Off Worlds, and decided to adjust the weapon type of it as the Convergence Bulldog wasn't needed for landing a real bullet on an enemy anymore. We chose the Striker Sidekick configuration for its red reticle range, zoom multiplier and default weapon skin. The ammo adjustment change accounts for the Striker Sidekick's different ammo capacity to the Convergence Bulldog.