feat: Add 37 Spike Off Ordo 'Mal - Added new Spike Grenade shooting weapon to the sandbox
We decided to add a Spike Grenade projectile weapon back after receiving a new node in the [Spring Update 2025](https://support.halowaypoint.com/hc/en-us/articles/34335369079700-Halo-Infinite-Spring-Update-2025-Patch-Notes#h_01JJQBZ4QWFM2D79NK0MQKSD39) for attributing scripted damage to a source object, and finding a way to consistently gather the Spike Grenade projectile again. These were both the main reasons why the [Spike Hydra was initially removed](https://github.com/The-Scripters-Guild/Warzone/pull/41).

As we now already have four Hydra variants, and we lost two Cindershot variants following the Fuel Rod SPNKr addition to the game, we had open slots for Cindershot variants and decided to convert the Spike Hydra to be a Cindershot variant. The Cindershot also allows for zooming, which is a welcome feature for this weapon.