feat: Add Carbine variants; replace #8 Removed weapon: - [8] Tovarus Artifice, REQ 7
Added weapons:
- [8] Scions Vision, REQ 7
- [45] Tactical Carbine, REQ 2
- [118] Vestige Carbine, REQ 4

**Tovarus Artifice Removal**
The [8] Tovarus Artifice was removed because it was basically a forcefully created variant of the Bandit, and it felt out of place.

**Vestige Carbine Addition**
The new Vestige Carbine has been added to the sandbox at REQ 4 after evaluating its efficiency. We believe this weapon will receive a nerf in the near future as it overpowers all other loadout weapons that it's currently being compared against by the developers as apparent from its use in official modes as a loadout weapon, and its class being Tactical Rifle.

**Tactical Carbine Addition**
After realizing the default Vestige Carbine was too powerful, we've designed a weaker version of it to fill a vacancy of a long-range weapon in REQ tier 2. This Carbine variant has notieable bullet spread, takes one more shot to kill and has much lower red reticle range.

**Scions Vision Addition**
As a replacement for the [8] Tovarus Artifice, we've designed a powerful variant of the Carbine that shoots in bursts of 3 shots, and can be fired very quickly. The weapon can kill a spartan in two bursts to the head, and is moslty effective in medium-range combat.
A very close competitor to this weapon is the Stalker Rifle Ultra, but the difference is the Stalker is more of a precision weapon for long-range while the Scions Vision is more of a run-and-gun type weapon.