feat: Adjust upgradedSprinting & upgradedWalking - Trait Set: upgradedSprinting   - Sprint Speed     - Top Speed Scalar: 1.25 → **1.15**
We've adjusted the Upgraded Sprinting Ability Boost to be less effective as it was too close to the speed of the Speed Boost powerup.

- Trait Set: upgradedWalking
  - Movement Speed: 1.13 → **1.20**
    - Sprint Speed
    - Top Speed Scalar: 1.00 → **0.90**

We've increased the movement speed of the Upgraded Walking Ability Boost due to the high jump feeling off when the walking speed was too slow in relation. The walking speed can now be compared to the normal running speed. Due to this, we've also compensated for the normally increased run speed to make the Ability Boost not have faster sprint.