feat: Gungoose, Warthog, Plasma Hog cost update - Adjusted vehicle costs:   - [3] Gungoose: 2 → 3   - [8] Warthog: 4 → 3   - [15] Plasma Warthog: 6 → 5
We've made adjustments to some vehicle costs based on feedback about them being used too much or too little.
The Gungoose seemed to be frequently picked at tier 2, and we've noticed it's a very strong vehicle for that tier especially against lower tier AI bosses. The Gungoose cost was increased to 3 so it doesn't become a best-in-class at tier 2.
The Warthog and Plasma Warthog were not seeing enough use so they were each dropped down a tier to see if players would start using them more. The vehicle requires two players to use, which is ultimately a large drawback to more of the casual audience who would rather play on their own.