feat: Loadout weapon tier overhaul; chart update - Overhauled loadout weapon levels:   - **Old**   - 1: [101] MA40 Assault Rifle, [105] Mk50 Sidekick   - 2: [1] Marksman Sidekick   - 3: [3] Hardlight Assault Rifle, [92] Bandit Evo   - 4: [102] MA5K Avenger, [99] M392 Bandit   - 5: [115] VK78 Commando, [4] Hardlight Avenger   - 6: [91] BR75, [9] Hardlight Commando   - 7: [5] Hardlight Battle Rifle   - **New**   - 1: [101] MA40 Assault Rifle, [105] Mk50 Sidekick   - 2: [1] Marksman Sidekick   - 3: [3] Hardlight Assault Rifle, [115] VK78 Commando   - 4: [9] Hardlight Commando   - 5: [102] MA5K Avenger   - 6: [4] Hardlight Avenger, [99] M392 Bandit   - 7: [91] BR75   - 8: [5] Hardlight Battle Rifle   - 9: [92] Bandit Evo
- Updated charts to have items in their own categories.

Loadout weapons have seen a major change in their level unlock tiers. We had been turning a blind eye on the loadout weapons for a long time, but playtester Kimbo Smiff pointed out that players were mostly using the Bandit Evo and nothing else after it unlocked.
The ranked gameplay expertise from Kimbo helped us realize that we had not been informed about the preferences of these weapons in a long time, and have now adjusted the order that the loadout weapons are unlocked in to more closely match how players would rank them in order.
This change along with adding 9 total Levels to the progression allows for enough time for each loadout weapon to be used during a match, and the progression and anticipation for the next loadout weapon is much more noticeable.