feat: Convergence Bulldog & Unbound PP cost incr. Increased cost of weapons: - [65] Convergence Bulldog: 5 → 6. - [83] Unbound Plasma Pistol: 5 → 6.
We've increased the cost of the Convergence Bulldog as it was way too consistent at killing players at short to medium ranges. This change also comes with decreasing the Valor Of Dinh damage to keep that weapon in tier 5, further displaying the gap between these weapons.
The Unbound Plasma Pistol received a cost increase due to its ease of use at getting instant kills with the charge shot if it lands on the head, which is quite easy to do due to the strong tracking of the weapon. At tier 5 the weapon felt a bit too easy to obtain for such easy kills.