feat: All Energy Swords cost increase Increased cost of weapons: - [96] Energy Sword: 3 → 4. - [35] Rogue Juggernaut: 4 → 5. - [68] Elite Bloodblade: 4 → 5. - [67] Duelist Energy Sword: 5 → 6. - [36] Demon: 6 → 7.
The cost of all Energy Sword variants have been increased as they all felt a bit too powerful for their tier in their own ways.
The regular Energy Sword was overpowering a CQS48 Bulldog at tier 3, the Duelist Energy Sword was too consistent at short-range combat that it overpowered the shotgun offerings, and Demon was too quirky and powerful of a weapon with its infinite camouflage to be kept at tier 6.
The Rogue Juggernaut and Elite Bloodblade tiers were increased to conform with the rest of the increase, but we will have to monitor them, especially the Elite Bloodblade to see if they still fit in the higher tier.