feat: Valor Of Dinh & Tactical Carbine damage adj. Adjusted weapon damage of: - [12] Valor of Dinh: 2.90 → 2.33 - [45] Tactical Carbine: 1.20 → 1.60
The Valor Of Dinh was feeling too powerful at tier 5 for the consistent kills it was getting, but we didn't want to just increase its tier as it was also performing nearly at the same pace as the Convergence Bulldog. We've instead reduced the damage of the weapon to a state where the instant kill range is now the same as the Mutilator 2nd fire consistent range.
The Tactical Carbine has seen a damage increase as it was still feeling too weak compared to the other tier 2 offerings. This change makes the Tactical Carbine a 6-shot perfect kill, which is faster than the Vestige Carbine, but you're extremely unlikely to get a perfect kill with the immense shot leading required when using the weapon.
The weapon now kills in 9 body shots, which combined with the shot leading gimmick should balance out the weapon as a skillful, but rewarding weapon at tier 2.