feat: Valkyrie & Light of Doisac adjustments [13] Valkyrie: - Tier: 8 → 9 - Ammo: 4+12 → 4+10
The Valkyrie was becoming too strong at tier 8 along with its high ammo capacity where it was able to win basically any fight as well as quickly destroy vehicles. With these in mind, we've nerfed the Valkyrie and bumped it to tier 9 to bring int more in line with the rest of the sandbox.

[46] Light Of Doisac
- Tier: 6 → 7
- Weapon Damage: 2.60 → 2.84
- Ammo: 4+16 → 4+8

The Light Of Doisac was initially designed as a variant of the S7 Flexfire, as it was able to shoot quicker than a normal Sniper, but it never felt that good as it only had 4 shots in the magazine. We've overhauled how this weapon is used by increasing the damage so landing all three Mangler bolts on the body will kill, just like the Valkyrie.
The downside of this weapon compared to the Valkyrie is its effective range as the Mangler bolts spread out the farther they travel, so landing all three on the enemy becomes less likely. With this change of making the weapon able to one-shot kill to the body, we've also reduced its ammo count and bumped it up to tier 7, as that one shot kill to the body is very powerful.