feat: Elite Bloodblade ammo and tier adjustments [68] Elite Bloodblade - Tier: 5 → 3 - Ammo: 100% → 40%
The Elite Bloodblade has been a weapon that nobody wants to purchase as players see it as just a normal Energy Sword that they don't know the benefits of. Due to this weapon being an outlier in how its communicated and it only having the benefit of dealing twice the amount of damage to vehicles compared to the normal Energy Sword, we've deviated from our rule of not adjusting default weapons by adjusting the ammo count of the weapon to 40%, which allows us to drop its tier significantly.
The ammo was dropped to 40% meaning that players can land three swings with it before the charge runs out. This allows players to kill three players, or quickly bring a Razorback to the doomsday state. The tier was dropped to 3 following the ammo change, so the weapon can be used at a low investment to gain a quick advantage if used effectively.