feat: Gamma Shot tier decrease [50] Gamma Shot - Points: 9 → 8
The Gamma Shot was dropped to tier 8 as it was being majorly overpowered by the Valkyrie earlier in tier 8 (now increased to 9), but it was also not living up to the standards of a tier 9 weapon. The weapon still restricts player movement capabilities, and players need to be in medium range to benefit from the tracking Cindershot shots which can instantly kill a player.