feat: Sentry Of Writh Kul movement speed decrease [34] Sentry Of Writh Kul - Movement Speed With Turret: 1.55 → 1.42
As pointed out by player "Precellence", moving diagonally with the Sentry Of Writh Kul could cause the player's partial camouflage to disappear due to the player moving too fast. We've now accounted for this by reducing the movement speed of the weapon to be just within the limits of how fast a player can move while still retaining the desired camouflage amount.