feat: M392 Bandit level decrease [99] M392 Bandit - Level: 6 → 5
Following feedback from player "Precellence", we've reduced the unlock level of the M392 Bandit by a tier so that players can use it earlier. We will monitor whether this change causes players to not switch away from the M392 Bandit to the higher-level offerings like the BR75 or Bandit Evo when they are unlocked. If players overwhelmingly stick with the M392 Bandit as their main weapon as its unlocked, it will get moved back up a tier.