feat: BR75 Breacher & Rapidfire Pulse tier incr. [62] BR75 Breacher - Points: 4 → 5
The BR75 Breacher was discovered to be very powerful at tier 4 due to its fast and consistent kill times and its high ammo count. With these in mind, we've bumped the weapon to tier 5.

[75] Rapidfire Pulse Carbine
- Points: 3 → 4

The Rapidfire Pulse Carbine was discovered to be an extreme improvement to the normal Pulse Carbine at just one tier lower. Landing one burst of 5 shots on a player would cause their shield to break, which isn't too difficult seeing how strong the tracking on the shots is for the weapon. The weapon can also be fired very quickly so missing a burst isn't a big deal. With these realizations in mind, we've bumped the weapon to tier 4.