feat: Tactical Carbine → Kig-Yar Carbine [45] Tactical Carbine → Kig-Yar Carbine - Weapon Type: Vestige Carbine + MA40 Longshot → Vestige Carbine + Pinpoint Needler   - Weapon Damage: 1.60 → - - Ammo Adjustment: -43.5 → 23   - 25+40 → 30+50
The Tactical Carbine has gotten an overhaul as we were still not happy with how it was performing. This overhaul makes the weapon shoot needles that require 8 shots to land on target for the supercombine effect which will kill a Spartan. The range of the weapon is slightly longer than a regular Needler, but it still loses out on short-range combat against the Needler due to the much slower fire rate.
The Kig-Yar Carbine is useful in mid-range when you want to suprise-attack an enemy. The weapon can also be creatively used to track players behind obstacles due to the lock-on persisting for a long time. The lock-on feature can also be used to curve the needles past obstacles to surprise the enemy and land shots on them while in cover.