feat: Hardlight BR, Commando, Avenger buffs The Hardlight loadout weapon variants were all having consistency issues. They already have a lot of drawbacks applied to them, and it seemed like players didnt value them enough compared to just running the normal versions of said weapons.
[5] Hardlight Battle Rifle
- Weapon Damage: 1.29 → 1.55

The Hardlight Battle Rifle was suffering from inconsistent 4-burst kills, so we've increased its damage where landing a 4-burst kill is a lot more reliable and consistent overall without removing the need to land 4 bursts to kill.

[9] Hardlight Commando
- Weapon Damage: 1.72 → 1.75

The Hardlight Commando has had its body shot Time To Kill lowered to increase the ease of use of the weapon and make any moments where a headshot that would have killed on a regular VK Commando not happening be less important; bridging the gap of consistency between the weapons to be more similar.

[4] Hardlight Avenger:
- Reload Speed: 0.75 → 1.00
- Ammo: 80+80 → 80+160

The Hardlight Avenger reload speed and ammo count have been adjusted to make the weapon the best at dealing quick damage to vehicles, as the weapon is a high-tier unlock and players naturally assume the weapon is good against vehicles after realizing that Hardlight damage is effetive against vehicles.