feat: New weapon: [55] Brute Carbine [55] Brute Carbine - Weapon Type: Vestige Carbine + Stalker Rifle Ultra - Trait Set: 55   - Weapon Damage: 0.30 - Config: Combo - Ammo Adjustment: -76   - 20+40 - Point Cost: 2 - Level: 2 - Notes: Good for landing consistent shots at long range.
The Brute Carbine was a weapon we came up with from still not being fully content with the Tactical Carbine overhaul that turned into the Kig-Yar Carbine. We wanted a weapon at tier 2 that could fill the gap of a cheap long-range weapon, but somehow still not make it too easy to use.
We finally came to a conclusion to use the Stalker Rifle Ultra config on the Carbine, which provided extreme long-range capability with a 123-meter red-reticle range and a high magnetism strength when in sight to make landing shots at long range easy. The weapon was made to kill in 8 headshots and 14 bodyshots, which was found to be a good balance seeing how much slower the Stalker Rifle Ultra config makes the Carbine shoot.