feat: Impact Commando tier decrease [69] Impact Commando - Points: 6 → 5
We've dropped the tier of the Impact Commando to 5 after re-evaluating the ease-of-use for the weapon. The weapon is notoriously inconsistent to use, making it disliked by many, and even if the user is knowledgeable on how to use it effectively by firing it in short 6-shot bursts, its still never pinpoint accurate. All these factors combined has made us drop the weapon down a tier to see if it performs better in tier 5.