feat: Decaying Charge adjustments [21] Decaying Charge - Weapon Type: Disruptor + Diminisher of Hope - Movement Speed: 1.10 → 1.00 - Ammo Adjustment: None → -45   - 12+36 → 12+20 - Point Cost: 4 → 5 - Level: 4 → 5
The Decaying Charge was found to be an overlooked and powerful weapon at tier 4 as the user could win most fights by landing 4 shots on target and waiting for the shock supercombine to eventually kill the enemy. This gimmick itself isn't bad, but the weapon being so cheap and having such high ammo capacity made it be too powerful, so we've bumped it up a tier and reduced its ammo count to try and alleviate its effiveness. We've also removed its movement speed buff as a further nerf.