feat: Various vehicle adjustments; mostly damage - Adjusted vehicle configs:   - [12] Mid Ghost: Health to 60% on spawn   - [13] Extermination Gungoose: Damage: 1.00 → 1.50 for driver   - [16] Phantom Wasp: Damage: 1.00 → 1.50 for driver, 5 sec health regen   - [17] Banishing Wasp → Moonlight Wasp   - [19] Dragon: Damage: 1.00 → 2.00 for driver, 5 sec health regen
We've added a passive health regeneration to some vehicles which heals the vehicle a percentage every interval. This may not seem like much, but it works in tandem with a native vehicle healing feature that heals every vehicle up to a regeneration threshold such as 80% health after usually 5-10 seconds of receiving no damage. If this passive health regeneration bumps the vehicle health to e.g. 81% and the vehicle has a regeneration threshold above it such as 100%, the vehicle will quickly heal to that threshold. This will essentially fully heal the vehicle over time. More info on vehicle regeneration thresholds: https://docs.google.com/spreadsheets/d/1YAw5lELFN_AUtxDYw4veAyjlvrkxY0Y5oNXEJ0P26YE/edit?gid=1656781007

The Mid Ghost was seen as potentially a bit too powerful for tier 3 as it's a single-manned vehicle that is agile and can cause good damage if used cleverly, so we've lowered its starting health to 60% to make it weaker in that regard.

The Extermination Gungoose wasn't seeing as consistent one-shot Ghost kills as we'd like to which was a gimmick of the vehicle so we've upped its damage to 1.50 to make these kills more likey.

The Phantom Wasp damage was increased to 1.50 as we were seeing it slightly underperforming for tier 8 vehicle expectations. This damage increase helps in getting more reliable infantry kills and also deals more damage using the rockets. The vehicle now features passive health regeneration.

The Banishing Wasp was overhauled to be the Moonlight Wasp. The weapon was changed to a Backdraft Cindershot and the damage was increased. The vehicle now features passive health regeneration.

The Dragon damage was increased to 2.00 as the weapon damage for the tier 9 vehicle was underperforming and not meeting expectations. The vehicle now features passive health regeneration.