feat: Nerf Upgraded Walking [5] Upgraded Walking: - Movement Speed: 1.20 → 1.10 - Jump Height: 1.25 → 1.15 - Sprint Speed, Top Speed Scalar: 0.90 → 1.00
We've nerfed the Upgraded Walking Ability Boost after seeing it was becoming a best-in-class pick for players. The movement speed was similar to the normal sprinting speed and the jump height was a bit too high to feel like a natural buff. We've decreased both of these as well as returned the sprint speed back to normal to compensate for the movement speed reduction, ensuring that the sprinting speed is still faster than normal sprint.