feat: Brute Carbine, Overloaded Pulse damage incr. [55] Brute Carbine - Weapon Damage: 0.30 → 0.35 The Brute Carbine was feeling too underpowered and inconsistent at getting kills for a tier 2 weapon so we've increased its damage to require one less headshot to kill a player. This should help with kill consistency and expectations from a tier 2 precision weapon.
[15] Overloaded Pulse Carbine:
- Weapon Damage: 1.70 → 3.00
- VFX: VIP → Infected Beta
The Overloaded Pulse Carbine damage was initially at 3.00, but was dropped to 1.70 after a session in a bot testing range which determined the gun too easy to use. After getting experience with other weapons in tier 7 during playtests, the kill consistency of this weapon didn't suit the other offerings as much as expected.
Also the weapon underperformed in destroying vehicles compared to the [33] Guardian Of Sanghelios at tier 6 due to the Overloaded Pulse Carbine's much lower ammo capacity. For these reasons, we've increased the weapon damage to the maximum 3.00 which now makes it the fastest weapon at dealing plasma damage at range in a short period. We've also adjusted the VFX of the weapon to the Infected Beta, which suits the threat level of the weapon better than the VIP effect.