Add vehicles: ONI Wasp, Kig-Yar Warthog (Sandbox 0.22.5), (Radial 0.6.4)
Added new vehicles:
- [25] ONI Wasp (Tier 7)
- [26] Kig-Yar Warthog (Tier 3)

The ONI Wasp is a variant of the Wasp that shoots Rocket Hog rockets. It has a unique effect of being able to displace units and vehicles with the explosion of the rocket, but it can also make actually getting kills difficult. It also has no secondary lock-on rocket to compensate for the rocket-firing primary.
The ONI Wasp is placed at tier 7 along with the normal Wasp since it's not necessarily and upgrade, but more of a different playstyle.

The Kig-Yar Warthog is our Halo 5 Needle Hog counterpart. Just in Halo Infinite the best we can do is a Warthog that shoots the slow-moving Pinpoint Needler shots instead of Needler Rifle shots like in Halo 5 so this vehicle has a different playstyle and kill potential.
The Needler shots travel slow, making landing them on an enemy be more difficult than faster moving projectiles. It takes 8 shots to supercombine and kill an enemy, but it can still be quite difficult to land all 8 in quick succession. The Needler shots also do very little damage to vehicles so this is strictly an infantry-only Warthog.
The Kig-Yar Warthog is placed at tier 3 due to these limitations of the Needler projectile and the lack of vehicle damage.