Adj. Carrier Falcon health, Kig-Yar Hog Shot Cad. (Sandbox 0.22.6)
Adjusted vehicles:
- [23] Carrier Falcon
  - Health Adjustment: 100% → **60%**
- [26] Kig-Yar Warthog
  - Shot Cadence: 0.0666 s → **0.00 s**

The Carrier Falcon health was adjusted to 60% since the Falcon natively has 1500 health, which is very high compared to the Banshee at 880 and Wasp at 840. A 60% scalar brings the initial health to 900, which is much more reasonable for a low-tier aerial vehicle.
Players can still choose to heal the Carrier Falcon to 1500 health with a Repair Field.