Weapon damage, ammo and tier adjustments (Sandbox 0.22.8), (Radial 0.6.6)
Adjusted weapons:
- [1] Marksman Sidekick
  - Weapon Damage: 0.66 → **0.71**
- [10] ONI Commando
  - Weapon Damage: 0.45 → **0.40**
  - Tier: 3 → **2**
- [61] Arcane Sentinel Beam:
  - Ammo: 80+80 → **80+160**
- [118] Vestige Carbine
  - Tier: 4 → **3**

The Marksman Sidekick was given more damage per bullet as getting kills with it was feeling too inconsistent and unrewarding.

The ONI Commando damage was decreased and its tier dropped from 3 → 2. The weapon is difficult to use, and it fits much better in tier 2 as a long-range option with a slight nerf to damage next to the carbine variants.

The Arcane Sentinel Beam ammo was increased by 80 based on feedback from player 'Precellence' about it losing the ammo too quickly. While we try to not adjust the ammo of default weapons, we are making an exception here as the weapon is fun to use, but it's limited by the ammo capacity—often being enough for just 4-6 kills.

The Vestige Carbine has seen a tier decrease from 4 → 3 due to it being not powerful enough for a tier 4 weapon status. It's now a strong tier 3, but there are other strong long-range weapons in the same tier so it doesn't feel like an outlier.