Valkyrie ammo, Kig-Yar Carbine damage etc. (Sandbox 0.22.9)
Adjusted weapons:
- [13] Valkyrie
  - Ammo: 4+10 → **4+4**
- [43] Banish Of Balaho
  - Ammo: 10+0 → **14+0**
- [45] Kig-Yar Carbine
  - Weapon Damage: 1.00 → **2.80**
- [47] Scorpion Tail
  - Ammo: 4% (6 shots) → **6% (9 shots)**

The Valkyrie (S7 Sniper Rifle) ammo has been decreased as the effectiveness of the high-tier sniper has been abundantly observed in Warzone Ultra matches where players gather points at an increased rate, and so can constantly purchase the Valkyrie.
This has exposed the high vehicle-killing capability of the weapon where it can take out tier 9 Wasps and Banshees in just two hits at any range. It's a difficult weapon to balance and we have thought about removing it from the sandbox due to how different vehicle health is in Halo Infinite compared to Halo 5, but we are still keeping it and just doing big changes to see what works and what doesn't.
Reducing the ammo to from 4+10 → 4+4 is a big reduction, but we want to test the limits in how to balance the weapon and when do players start using weapons that are intended for only anti-vehicle or anti-infantry, instead of always opting for the best-in-class weapon that can do both.

The Banish of Balaho (Skewer) ammo has been increased from 10+0 → 14+0 to give it more value and damage potential. The weapon is mainly for anti-vehicle, but the shots are harder to hit than a hitscan sniper rifle.
This change is done to see if players start using the weapon more over the Valkyrie for anti-vehicle.

The Kig-Yar Carbine (Vestige Carbine) damage was increased from 1.00 → 2.80 to make the individual needles that don't result in a supercombine still deal damage to the enemy. We noticed it was frustrating when you didn't land the 8 required needles to supercombine and kill an enemy, and then realizing you also had done only a very small amount of damage to the enemy.
Increasing the damage means that even if you don't get the supercombine, at least the enemy is weaker and can be more easily killed following your initial gunfight as long as they don't get a chance to regenerate their shields and health in between.

The Scorpion Tail ammo was increased from 6 shots → 9 shots to bump up its value as a tier 9 weapon in comparison to mainly the Valkyrie. The Scorpion Tail still gives the player slow movement speed, and its projectile has a noticeable travel time, so while it deals a strong blow, you need to hit that in the first place.